Split out integration test plumbing (#2594) (#2300)
