+++ integration-testing/Cargo.toml
@@ -36,12 +36,15 @@ arrow-flight = { path = "../arrow-flight", default-features = false }
 async-trait = { version = "0.1.41", default-features = false }
 clap = { version = "3", default-features = false, features = ["std", "derive"] }
 futures = { version = "0.3", default-features = false }
-hex = { version = "0.4", default-features = false }
+hex = { version = "0.4", default-features = false, features = ["std"] }
 prost = { version = "0.11", default-features = false }
-serde = { version = "1.0", default-features = false, features = ["rc"] }
-serde_derive = { version = "1.0", default-features = false }
+serde = { version = "1.0", default-features = false, features = ["rc", "derive"] }
 serde_json = { version = "1.0", default-features = false, features = ["std"] }
 tokio = { version = "1.0", default-features = false }
 tonic = { version = "0.8", default-features = false }
 tracing-subscriber = { version = "0.3.1", default-features = false, features = ["fmt"], optional = true }
 num = { version = "0.4", default-features = false, features = ["std"] }
+flate2 = { version = "1", default-features = false, features = ["rust_backend"] }
+
+[dev-dependencies]
+tempfile = { version = "3", default-features = false }
